requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
jinja2
python-multipart
requests
pydantic
openai
